(PC-30240) refactor(ATT): Rename and test function
